rebackend/requirements.txt
# ── Web Framework ──────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# ── Configuration ──────────────────────────────────────────────────────────────
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1

# ── LLM / AI ───────────────────────────────────────────────────────────────────
groq==0.9.0

# ── Vector Store ───────────────────────────────────────────────────────────────
# chromadb 0.5.x pulls numpy>=2.0 which removed np.float_ → AttributeError.
# 0.4.24 is the last release with stable numpy<2 compat.
chromadb==0.4.24
numpy<2.0

# ── Graph ──────────────────────────────────────────────────────────────────────
networkx==3.3

# ── Git ────────────────────────────────────────────────────────────────────────
gitpython==3.1.43

# ── HTTP ───────────────────────────────────────────────────────────────────────
httpx==0.27.0
requests==2.32.2

# ── Utils ──────────────────────────────────────────────────────────────────────
tiktoken==0.7.0
aiofiles==23.2.1
